satellite wake command
